Even more mutation coverage
